Update macro generated
Improve the macro generated so it checks if the mob was actually targeted before assigning a raid icon. Without this improvement, the raid icon is affected regardless of a successful targeting, making you set raid icon on a random mob that is attacking you if you're running across the land with a lot of them behind you.